+++ Semana_3/githubrepo/src/Styles/global.ts
@@ -23,7 +23,7 @@ body,input,button{
     padding: 40px 20px
 }
 
-button{
+button,a{
     cursor:pointer
 }
 `;